Background code added
This code adds the background feature on the code, it uses an API that randomly GET and updates the background image of the webpage with a random nature photo from Unsplash and displays the name of the photographer. If there's an error, it falls back to a default background image and author.